+++ src/main/java/br/com/zalf/prolog/webservice/dashboard/components/charts/line/HorizontalLimitLine.java
@@ -1,5 +1,6 @@
 package br.com.zalf.prolog.webservice.dashboard.components.charts.line;
 
+import br.com.zalf.prolog.webservice.dashboard.Color;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -11,10 +12,13 @@ public final class HorizontalLimitLine {
     private final double posicaoY;
     @NotNull
     private final String descricao;
+    @NotNull
+    private final Color cor;
 
-    public HorizontalLimitLine(final double posicaoY, @NotNull final String descricao) {
+    public HorizontalLimitLine(final double posicaoY, @NotNull final String descricao, @NotNull final Color cor) {
         this.posicaoY = posicaoY;
         this.descricao = descricao;
+        this.cor = cor;
     }
 
     public double getPosicaoY() {
@@ -24,5 +28,10 @@ public final class HorizontalLimitLine {
     @NotNull
     public String getDescricao() {
         return descricao;
+    }
+
+    @NotNull
+    public Color getCor() {
+        return cor;
     }
 }